NumPy stubs: Fix `rand` DType

+++ stubs/numpy/random.pyi
@@ -1,6 +1,6 @@
 from pyre_extensions import TypeVarTuple, Unpack
 from typing import Any, TypeVar, Type, Tuple, overload
-from . import ndarray, float
+from . import ndarray, int64, float32, float64
 
 Ts = TypeVarTuple("Ts")
 N = TypeVar("N", bound=int)
@@ -8,7 +8,7 @@ DType = TypeVar("DType")
 
 Number = Union[int, float]
 
-def randn(*args: Unpack[Ts]) -> ndarray[float, Unpack[Ts]]: ...
+def randn(*args: Unpack[Ts]) -> ndarray[float64, Unpack[Ts]]: ...
 @overload
 def randint(
     low: int, high: int = ..., size: Tuple[Unpack[Ts]] = ..., dtype: Type[DType] = ...
